+++ XpringKit/PayID/PayIDClient.swift
@@ -17,14 +17,13 @@ public class PayIDClient {
   /// The network this PayID client resolves on.
   private let network: String
 
-  // The queue to execute callbacks on.
-  private let completionQueue: DispatchQueue
+  // A queue to perform networking on.
+  private let networkQueue = DispatchQueue(label: "io.xpring.PayIDClient", qos: .userInitiated)
 
   /// Initialize a new PayID client.
   ///
   /// - Parameters:
   ///   - network: The network that addresses will be resolved on.
-  ///   - completionQueue: The queue that callback blocks will be executed on.  Defaults to the main queue.
   ///
   /// - Note: Networks in this constructor take the form of an asset and an optional network (<asset>-<network>).
   /// For instance:
@@ -36,29 +35,6 @@ public class PayIDClient {
   //  TODO: Link a canonical list at payid.org when available.
   public init(network: String, completionQueue: DispatchQueue = DispatchQueue.main) {
     self.network = network
-    self.completionQueue = completionQueue
-  }
-
-  public func address(for payID: String) throws -> Result<CryptoAddressDetails, PayIDError> {
-    var retrunResult: Result<CryptoAddressDetails, PayIDError>!
-
-    // TODO: Use a semaphore.
-    let group = DispatchGroup()
-    group.enter()
-
-    let queue = DispatchQueue.global(qos: .userInitiated)
-
-   queue.async {
-    print("Queue Executing")
-      self.address(for: payID, completionQueue: queue) { result in
-        print("Queue called back")
-        retrunResult = result
-        group.leave()
-      }
-   }
-    group.wait()
-
-    return retrunResult
   }
 
   /// Resolve the given PayID to an address.
@@ -66,10 +42,34 @@ public class PayIDClient {
   /// - Parameter payID: The PayID to resolve for an address.
   /// - Parameter completion: A closure called with the result of the operation.
   /// - Returns: An address representing the given PayID.
-  // TODO(keefertaylor): Make this API synchronous to mirror functionality provided by ILP / XRP.
+  public func address(for payID: String) throws -> Result<CryptoAddressDetails, PayIDError> {
+    // Assign a bogus value. This will get overwritten when the asynchronous call is made.
+    var result: Result<CryptoAddressDetails, PayIDError> = .failure(.unknown(error: "Unknown error."))
+
+    // Use a semaphore to block and wait for the asynchronous call to complete.
+    // Capture the resolved data in result.
+    let semaphore = DispatchSemaphore(value: 0)
+    self.address(for: payID) { resolvedResult in
+      // Capture the result of the call.
+      result = resolvedResult
+
+      // Signal to the semaphore to unblock the thread.
+      semaphore.signal()
+    }
+
+    // Wait for networking to complete.
+    semaphore.wait()
+
+    return result
+  }
+
+  /// Resolve the given PayID to an address.
+  ///
+  /// - Parameter payID: The PayID to resolve for an address.
+  /// - Parameter completion: A closure called with the result of the operation.
+  /// - Returns: An address representing the given PayID.
   public func address(
     for payID: String,
-    completionQueue: DispatchQueue = DispatchQueue.main,
     completion: @escaping (Result<CryptoAddressDetails, PayIDError>) -> Void
   ) {
     guard let payIDComponents = PayIDUtils.parse(payID: payID) else {
@@ -88,7 +88,7 @@ public class PayIDClient {
 
     let request = API.ResolvePayID.Request(path: path)
 
-    client.makeRequest(request, completionQueue: completionQueue) { apiResponse in
+    client.makeRequest(request, completionQueue: self.networkQueue) { apiResponse in
       switch apiResponse.result {
       case .success(let response):
         switch response {